Initial refactor of ServerModel to TypeScript
