+++ src/utils.ts
@@ -1,0 +1,14 @@
+const isObject = (value: any): boolean => {
+    var objType = Object.prototype.toString.call(value);
+    return objType === '[object Object]' || objType === '[object Error]';
+};
+
+const parseNumber = (value: string | number): number => {
+    if (isNaN(value as number) || !isFinite(value as number)) {
+        return 0;
+    }
+    var floatValue = parseFloat(value as string);
+    return isNaN(floatValue) ? 0 : floatValue;
+};
+
+export { isObject, parseNumber };
